cope with case insensitive filesystems when branding MacOS eclipse.ini -  Bug 325302

+++ org.eclipse.buckminster.pde/src/java/org/eclipse/buckminster/pde/tasks/BrandingIron.java
@@ -321,7 +321,7 @@ public class BrandingIron implements IXMLConstants {
 			ini = brandedIni;
 		} else {
 			if (ini.exists()) {
-				if (ini2.exists()) {
+				if (ini2.exists() && !ini2.equals(ini)) {
 					logger.warning("Found both %s and %s. Discarding the latter", ini.getAbsolutePath(), ini2.getAbsolutePath()); //$NON-NLS-1$
 					ini2.delete(); // This should not happen really
 				}
